Delete setup.cfg
