+++ biliup/plugins/bili_webup.py
@@ -19,7 +19,7 @@ import rsa
 import xml.etree.ElementTree as ET
 from requests.adapters import HTTPAdapter, Retry
 
-from biliup import config
+from ..engine import config
 from ..engine import Plugin
 from ..engine.upload import UploadBase, logger
