--- requirements.txt
@@ -1,12 +1,0 @@
-selenium>=3.8.1
-psutil>=5.4.6
-yt_dlp>=2022.3.8.2
-Pillow>=5.2.0
-aiohttp[speedups]
-Requests>=2.20.0
-PyYAML>=4.2b1
-streamlink>=0.14.2
-ykdl>=1.8.0
-rsa>=4.6
-aiohttp_cors
-aiohttp_basicauth_middleware
