--- setup.cfg
@@ -1,35 +1,0 @@
-[metadata]
-name = biliup
-version = attr: biliup.__version__
-author = ForgQi
-url = https://github.com/ForgQi/bilibiliupload
-description = stream download and upload
-long_description = file: README.md
-long_description_content_type = text/markdown
-license_files = LICENSE
-keywords = bilibili douyu huya
-
-[options]
-packages = find:
-python_requires = >= 3.7
-install_requires =
-  selenium >= 3.8.1
-  psutil >= 5.4.6
-  yt_dlp >= 2022.3.8.2
-  Pillow >= 5.2.0
-  aiohttp[speedups]
-  Requests >= 2.20.0
-  PyYAML >= 4.2b1
-  streamlink >= 0.14.2
-  ykdl >= 1.8.0
-  rsa >= 4.6
-
-[options.package_data]
-biliup = common/*.ini,public/*,public/build/*
-
-[options.packages.find]
-include = biliup*
-
-[options.entry_points]
-console_scripts =
-    biliup = biliup.__main__:arg_parser
